+++ CHANGELOG.md
@@ -16,6 +16,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `deletable' field on post
 - Admin field that for admin users
 - ability for admins to delete posts
+- ability to upload file at `/upload` with the name profilePicture
+- publicPath to config file to configure the directory for public files
 
 ### Removed
 
@@ -26,6 +28,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - changed the running behaviour to run in cluster threads via node.js cluster api
 - gql field userVote requires a userId
 - default findUser param limit to 20
+- only group admins can create group events
 
 ### Fixed